@@ -13,7 +13,7 @@ from starlette.formparsers import FormParser, MultiPartException, MultiPartParse
 from starlette.types import Message, Receive, Scope, Send
 
 if typing.TYPE_CHECKING:
-    from multipart.multipart import parse_options_header
+    from python_multipart.multipart import parse_options_header
 
     from starlette.applications import Starlette
     from starlette.routing import Router
@@ -230,7 +230,7 @@ class Request(HTTPConnection):
                     self._stream_consumed = True
                 if body:
                     yield body
-            elif message["type"] == "http.disconnect":
+            elif message["type"] == "http.disconnect":  # pragma: no branch
                 self._is_disconnected = True
                 raise ClientDisconnect()
         yield b""
@@ -250,12 +250,17 @@ class Request(HTTPConnection):
         return self._json
 
     async def _get_form(
-        self, *, max_files: int | float = 1000, max_fields: int | float = 1000, max_file_size: int | None
+        self,
+        *,
+        max_files: int | float = 1000,
+        max_fields: int | float = 1000,
+        max_part_size: int = 1024 * 1024,
+        max_file_spool_size: int = 1024 * 1024,
     ) -> FormData:
-        if self._form is None:
-            assert (
-                parse_options_header is not None
-            ), "The `python-multipart` library must be installed to use form parsing."
+        if self._form is None:  # pragma: no branch
+            assert parse_options_header is not None, (
+                "The `python-multipart` library must be installed to use form parsing."
+            )
             content_type_header = self.headers.get("Content-Type")
             content_type: bytes
             content_type, _ = parse_options_header(content_type_header)
@@ -266,7 +271,8 @@ class Request(HTTPConnection):
                         self.stream(),
                         max_files=max_files,
                         max_fields=max_fields,
-                        max_part_file_size=max_file_size,
+                        max_part_size=max_part_size,
+                        spool_max_size=max_file_spool_size,
                     )
                     self._form = await multipart_parser.parse()
                 except MultiPartException as exc:
@@ -281,10 +287,20 @@ class Request(HTTPConnection):
         return self._form
 
     def form(
-        self, *, max_files: int | float = 1000, max_fields: int | float = 1000, max_file_size: int | None = None
+        self,
+        *,
+        max_files: int | float = 1000,
+        max_fields: int | float = 1000,
+        max_part_size: int = 1024 * 1024,
+        max_file_spool_size: int = 1024 * 1024,
     ) -> AwaitableOrContextManager[FormData]:
         return AwaitableOrContextManagerWrapper(
-            self._get_form(max_files=max_files, max_fields=max_fields, max_file_size=max_file_size)
+            self._get_form(
+                max_files=max_files,
+                max_fields=max_fields,
+                max_part_size=max_part_size,
+                max_file_spool_size=max_file_spool_size,
+            )
         )
 
     async def close(self) -> None:
